Fix error handling of odb_object__alloc() in odb_read_1()
If `odb_object__alloc()` returns `NULL` (fails to allocate) it currently goes to `out`. If error is != 0 we will try to free raw.data, when raw couldn't be allocated. If error is 0, then we are silently swallowing the allocation error, and possibly causing access to a wrong `out` pointer. Let's return an error instead.