atexit: create skeleton
This provides a skeleton for registering rollback functions in case the caller
wants to centralise their error handling in an `atexit(3)`-like manner.

Subsystems within libgit2 will register themselves when they e.g. create
lockfiles. If the caller detects an error (potentially unrelated to the libgit2
operation) and wants to back out, they could call `git_atexit()` in order to
remove lockfiles which the library is holding.