Add revwalk filter API
This works similarly to the hide API. It allows the caller to specify
a callback to be invoked at optimal times during the revwalk
process in order to give the caller control over which commits to filter
from iteration. This is different from the hide API in that only the
commit in question will be potentially filtered.