futils: map read-only memory as PRIVATE
The function `git_futils_mmap_ro` handles mapping a memory segment as
read-only. While we do correctly use `GIT_PROT_READ` to mark it as
read-only, we pass in `GIT_MAP_SHARED` as flags to `mmap`. The
difference between shared and private mappings is that all writes to the
mapped region will change the underlying object for a shared mapping.
But as the region is read-only anyway, we should just map it as private.