posix: fix compilation error on including "unix/posix.h"
When including "unix/posix.h" directly, the compiler will produce an
error due to the variable `p_fsync__cnt` not being defined. The variable
is declared as extern in "posix.h", but as the platform-specific
POSIX-headers do not include that the definition is unknown.

Fix the issue by duplicating the declaration to "unix/posix.h".