config_file: remove unused list iteration macros
We currently provide a lot of macros for the `cvar_t` structure which
are never being used. In fact, the only macro we need is to access the
`next` pointer of `cvar_t`, which really does not require a macro at
all.

Remove all these macros and replace usage of `CVAR_LIST_NEXT(cvar)` with
`cvar->next`.