odb: free write stream in case of error
As soon as we have opened a write stream in `git_odb_open_wstream`, we
will never free it when erroring out of the function. This contradicts
our usual style, where out parameters will only ever be set and need to
be freed when the function returns success.

Convert the function to always free allocated streams in case an error
occurs.