odb: improve logic around unsupported writing in backends
In case no backend supports either `writestream` or `write`, we want to
raise an error to the caller and make him aware of this deficiency.
Right now, this logic depends on the fact that `error` is being
initialized to `-1`, which is rather hard to spot on first sight.

Move the error check out of the `error < 0` conditional and instead only
check whether the `writes` variable, which keeps track of the number of
backends supporting writes, has been incremented at all. This greatly
helps readability.