odb: prioritize write streams over fake write streams
When opening a write stream for the ODB, we loop through each backend
and try to either open it as a normal write stream or, in case
`writestream` has not been implented by the backend, a fake write
stream. The fake write stream simply wraps the backend's `write`
callback. This has the problem though that even if there is a backend
with write streams being supported, we will never use them as long as
there is another backend without a `writestream`, but a `write`
function. This is kind of an awkward interface to use, as it effectively
hinders you to use a lower-prioritized `writestream` only backend for
e.g. streaming out large objects as soon as you have another backend
with simple write support.

Rewrite the logic to make two passes over the list of backends. In the
first pass, we will for each of the backends try to use the
`writestream` function, if it exists. Only in the second pass will we
now try to open a fake write stream. This effectively gives backends
implementing `writestream` a higher priority as every other backend only
implementing `write`. As the most important purpose of using streams is
to help in writing huge objects, this is the right thing to do, as we do
not want to silently fall back on `write`.